UT: get_RowFromIndex_Typical IsMineUp_WhenFalse IsMineUp_WhenNoUpSquare

+++ MineGame_Lib/SquareFactory.h
@@ -18,15 +18,16 @@ public:
 	virtual vector<ISquare*> CreateSquares(MGame* game, MineField* mineField);
 	virtual vector<bool> GenerateMineFieldMap(const MineField& mineField);
 
-	static bool HasAdjacentMine( const vector<bool>& squareMap, const Size& fieldSize, size_t i );
-	static bool IsMineUp( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineDown( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineLeft( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineRight( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineUpRight( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineDownRight( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineUpLeft( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
-	static bool IsMineDownLeft( const vector<bool>& fieldMap, const Size& fieldSize, size_t i );
+	static bool HasAdjacentMine(const vector<bool>& squareMap, const Size& fieldSize, size_t i);
+
+	static bool IsMineUp		(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineDown		(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineLeft		(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineRight		(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineUpRight	(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineDownRight	(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineUpLeft	(const vector<bool>& fieldMap, const Size& fieldSize, int i);
+	static bool IsMineDownLeft	(const vector<bool>& fieldMap, const Size& fieldSize, int i);
 
 private:
 	Random _random;
